+++ packages/clients/meldemichel/src/types.ts
@@ -4,6 +4,7 @@ export interface MeldemichelCreateMapParams {
   containerId: string
   mode: keyof typeof MODE
   afmUrl?: string
+  stadtwaldActive?: boolean
   reportServiceId?: string
   configOverride?: object
 }