+++ packages/plugins/Draw/src/components/Draw.vue
@@ -17,6 +17,14 @@
         :change-callback="setDrawMode"
       ></RadioCard>
       <DrawOptions v-if="showDrawOptions" />
+      <RadioCard
+        v-if="showMeasureOptions"
+        id="draw-measure"
+        title="common:plugins.draw.title.measureMode"
+        :initial-value="measureMode"
+        :values="selectableMeasureModes"
+        :change-callback="setMeasureMode"
+      ></RadioCard>
       <v-subheader v-if="showSizeSlider" class="align-end">{{
         $t('common:plugins.draw.label.textSize')
       }}</v-subheader>
@@ -61,16 +69,19 @@ export default Vue.extend({
   computed: {
     ...mapGetters(['hasSmallHeight', 'hasWindowSize']),
     ...mapGetters('plugin/draw', [
-      'mode',
       'drawMode',
-      'selectableDrawModes',
-      'selectableModes',
-      'textInput',
-      'showTextInput',
-      'selectedSize',
       'fontSizes',
-      'showSizeSlider',
+      'measureMode',
+      'mode',
+      'selectableDrawModes',
+      'selectableMeasureModes',
+      'selectableModes',
       'showDrawOptions',
+      'showMeasureOptions',
+      'selectedSize',
+      'showSizeSlider',
+      'showTextInput',
+      'textInput',
     ]),
     flexStyle(): string {
       return `flex-direction: ${
@@ -80,10 +91,11 @@ export default Vue.extend({
   },
   methods: {
     ...mapActions('plugin/draw', [
-      'setMode',
       'setDrawMode',
-      'setTextInput',
+      'setMeasureMode',
+      'setMode',
       'setSelectedSize',
+      'setTextInput',
     ]),
   },
 })